FRS/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.11.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
dataclasses==0.8
deepface==0.0.73
filelock==3.4.1
Flask==2.0.3
flatbuffers==1.12
gast==0.4.0
gdown==4.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.0.1
mmh3==3.0.0
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
pandas
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymilvus==2.0.0rc9
PySocks==1.7.1
pyspark==2.4.7
python-dateutil==2.8.2
pytz==2022.1
pywebhdfs
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.8
six==1.15.0
soupsieve==2.3.2.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.64.0
typing-extensions==3.7.4.3
ujson==4.3.0
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
